chore: gather utils into one file
