+++ src/api/fetchTotalDataByCountry.ts
@@ -1,7 +1,7 @@
 import axios from 'axios';
-import type { TotalData, countries } from '../types';
+import type { TotalData } from '../types';
 
-const fetchDataByCountry = async (country: countries): Promise<[TotalData, TotalData]> => {
+const fetchDataByCountry = async (country: string): Promise<[TotalData, TotalData]> => {
   const [response1, response2] = await Promise.all([
     axios.get<TotalData>(`countries/${country}?yesterday=false&strict=false`),
     axios.get<TotalData>(`countries/${country}?yesterday=true&strict=false`),
